still preparing for new consensus

+++ blockchain/params/params.go
@@ -7,7 +7,7 @@ import (
 
 type LoanParams struct {
 	InterestRate     uint64 `json:"InterestRate"`     // basis points, e.g. 125 represents 1.25%
-	Maturity         uint32 `json:"Maturity"`         // in number of blocks
+	Maturity         uint64 `json:"Maturity"`         // in number of blocks
 	LiquidationStart uint64 `json:"LiquidationStart"` // ratio between collateral and debt to start auto-liquidation, stored in basis points
 }
