+++ database/lvdb/loan.go
@@ -76,12 +76,12 @@ func (db *db) GetLoanTxs(loanID []byte) ([][]byte, error) {
 	return results, nil
 }
 
-func (db *db) StoreLoanPayment(loanID []byte, principle, interest uint64, deadline uint32) error {
+func (db *db) StoreLoanPayment(loanID []byte, principle, interest uint64, deadline uint64) error {
 	// TODO(@0xbunyip): implement
 	return nil
 }
 
-func (db *db) GetLoanPayment(loanID []byte) (uint64, uint64, uint32, error) {
+func (db *db) GetLoanPayment(loanID []byte) (uint64, uint64, uint64, error) {
 	// TODO(@0xbunyip): implement
 	return 0, 0, 0, nil
 }
